Update bootrr commit id
This new revision includes some TPM tests which can be run now that
the required dependencies have been added to the baseline fragment.

It could also have been done by defining dependencies in the bootrr
package, but it may not be required to have those tools when not
running the TPM tests.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>